Moved class singleAPI to api module and `import api` below config check
